requirements.txt
Django==5.1.6
djangorestframework>=3.14.0
drf-yasg>=1.21.6
celery>=5.5.0
redis>=4.5.1
psycopg2-binary>=2.9.0
django-oauth-toolkit>=1.7.0
